+++ CHANGELOG.md
@@ -1,5 +1,41 @@
 # Changelog
 
+## 0.16.0 - 2024-03-01
+### Enhancements
+- Updated `StatusMsg` and made it public in preparation for releasing a status schema
+- Added `StatusAction`, `StatusReason`, `TradingEvent`, and `TriState` enums for use in
+  the status schema
+- Added `-t` and `--tsv` flags to DBN CLI to encode tab-separated values (TSV)
+- Added `delimiter` method to builders for `DynEncoder` and `CsvEncoder` to customize the
+  field delimiter character, allowing DBN to be encoded as tab-separated values (TSV)
+- Document cancellation safety for `AsyncRecordDecoder::decode_ref` (credit: @yongqli)
+- Added new publisher values for consolidated DBEQ.MAX
+- Added C FFI conversion functions from `ErrorMsgV1` to `ErrorMsg` and `SystemMsgV1`
+  to `SystemMsg`
+- Improved documentation for `side` field and `Side` enum
+- Upgraded `async-compression` to 0.4.6
+- Upgraded `strum` to 0.26
+
+### Breaking changes
+- Changed default for `VersionUpgradePolicy` to `Upgrade`
+- Changed default `upgrade_policy` for `DbnDecoder`, `AsyncDbnDecoder`, and Python
+  `DBNDecoder` to `Upgrade` so by default the primary record types can always be used
+- Changed fields of previously-hidden `StatusMsg` record type
+- Updated text serialization order of status schema to match other schemas
+- Changed text serialization `unit_of_measure_qty` to be affected by `pretty_px`. While
+  it's not a price, it uses the same fixed-price decimal format as other prices
+- Made `StatType` and  `VersionUpgradePolicy` non-exhaustive to allow future additions
+  without breaking changes
+- Renamed `_dummy` field in `ImbalanceMsg` and `StatMsg` to `_reserved`
+- Added `ts_out` parameter to `RecordDecoder` and `AsyncRecordDecoder`
+  `with_upgrade_policy` methods
+
+### Bug fixes
+- Fixed handling of `ts_out` when upgrading DBNv1 records to version 2
+- Added missing `StatType::Vwap` variant used in the ICE datasets
+- Fixed an issue with Python stub file distribution
+- Fixed missing handling of `ErrorMsgV1` and `SystemMsgV1` in `rtype` dispatch macros
+
 ## 0.15.1 - 2024-01-23
 
 ### Bug fixes
@@ -23,7 +59,7 @@
     to forget
 - Added `-s`/`--map-symbols` flag to CLI to create a `symbol` field in the output with
   the text symbol mapped from the instrument ID
-- Added `version` param to Python `Metadata` contructor choose between DBNv1 and DBNv2
+- Added `version` param to Python `Metadata` constructor choose between DBNv1 and DBNv2
 - Implemented `EncodeRecordTextExt` for `DynEncoder`
 - Implemented `Deserialize` and `Serialize` for all records and enums (with `serde`
   feature enabled). This allows serializing records with additional encodings not
@@ -113,7 +149,7 @@
   `dbn::AsyncRecordDecoder::decode`, and `dbn::AsyncRecordDecoder::decode_ref`
   cancellation safe. This makes them safe to use within a
   `tokio::select!`(https://docs.rs/tokio/latest/tokio/macro.select.html) statement
-- Added documention around cancellation safety for async APIs
+- Added documentation around cancellation safety for async APIs
 - Improved error messages for conversion errors
 - Added `TOB` flag to denote top-of-book messages
 - Added new publisher values in preparation for IFEU.IMPACT and NDEX.IMPACT datasets